Fix id parsing as int64

+++ src/aglet/DTO/GithubDTO.fs
@@ -5,7 +5,7 @@ open Domain
 
 type LabelInfoResponse = {
     [<JsonProperty("id")>]         
-    ID: int
+    ID: int64
     [<JsonProperty("node_id")>]     
     NodeId : string
     [<JsonProperty("url")>]
